+++ src/main/java/org/sonar/plugins/delphi/antlr/ast/node/ArrayIndicesNode.java
@@ -9,6 +9,10 @@ public final class ArrayIndicesNode extends DelphiNode {
     super(token);
   }
 
+  public ArrayIndicesNode(int tokenType) {
+    super(tokenType);
+  }
+
   @Override
   public <T> T accept(DelphiParserVisitor<T> visitor, T data) {
     return visitor.visit(this, data);