+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for obscure `(.` and `.)` tokens:
+  - These are now interpreted correctly by the lexer as square brackets.
+  - See: [The Future of the Delphi Compiler](https://web.archive.org/web/20210330140547/https://edn.embarcadero.com/article/39174)
+
 ## [0.29.0] - 2021-03-23
 
 ### Added